django_service_1/requirements/requirements.txt
django-cors-headers==4.7.0
django-redis==5.4.0
django-role-permissions==3.2.0
djangorestframework_simplejwt==5.5.0
google-generativeai==0.8.5
groq==0.23.1
gunicorn==23.0.0
openai==1.82.0
openpyxl==3.1.5
pipdeptree==2.26.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
reportlab==4.4.0
resend==2.10.0